Many bug fixes for benchmarks
Benchmarks now handle hexes and squares. The hex and square timings are a little
rough still. Also, fix the sound_delay in ants to consistently use the real value
of actual_frames, which is in turn correctly computed to be the total delay since
the last frame we ran, not the delay of the last single frame (which, if there
were any idle frames, will nearly always show as 1).